moved components and styling out of app.js, need to finish stly on other components later, going to try to get it to TS and add tests next before any more features or styling
