+++ styles/Styles.js
@@ -1,0 +1,20 @@
+import React from 'react';
+import { StyleSheet } from 'react-native';
+
+
+// Styles
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        backgroundColor: 'black',
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    button: {
+        color: 'green',
+        //backgroundColor: 'blue',
+        fontSize: 40,
+    }
+});
+
+export default styles;
